python/analysis/pyproject.toml
[project]
name = "agent-plot-analysis"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "numpy>=2.0",
  "tifffile>=2024.8.30",
  "pillow>=10.0",
  "pandas>=2.2",
]
